requirements.txt
kafka-python
requests
pandas
streamlit